Update examples to use latest APIs. Update READMEs. Delete duplicate examples
