+++ examples/gpt2/multipack_pipeline_gpt2.py
@@ -2,10 +2,9 @@ import json
 
 from termcolor import colored
 
-from forte.data.ontology import base_ontology
 from forte.multipack_pipeline import MultiPackPipeline
 
-from ft.onto.base_ontology import Sentence
+from ft.onto.base_ontology import Token, Sentence
 
 pl = MultiPackPipeline()
 
@@ -28,12 +27,12 @@ print(f'sentence_cnt: src{src_cnt}, tgt{tgt_cnt}, link_cnt{link_cnt}')
 with open("multipack_output.txt", "w+") as fout:
     input_pack = multipack.get_pack(input_pack_name)
     output_pack = multipack.get_pack(output_pack_name)
-    for context, gen_sentence in zip(input_pack.get(base_ontology.Sentence),
-                        output_pack.get(base_ontology.Sentence)):
+    for context, gen_sentence in zip(input_pack.get(Sentence),
+                        output_pack.get(Sentence)):
         print(colored("Initial Context", "green"), context.text)
         print(colored("Generated Sentence", "green"), gen_sentence.text)
         print("======================TAGS======================")
-        for token in output_pack.get(entry_type=base_ontology.Token,
+        for token in output_pack.get(entry_type=Token,
                                      range_annotation=gen_sentence):
             print(colored("Token", "red"), token.text,
                   colored("POS Tag", "red"), token.pos_tag)
