--- examples/notebooks/sample_config.yml
@@ -1,26 +1,0 @@
-# This provides a configuration for interence only
-
-# TODO: since we use pipeline.init_from_config(Hparams), we cannot set the PipelineType in the config
-# Unless we have an external code parsing this configuratoin and create the Pipeline itself.
-
-# PipelineType: SinglePackPipeline
-# Or maybe MultiPackTrainPipeline
-
-# Ontology:
-Reader:
-  type: CoNLL03Reader
-  kwargs: null
-
-Processors:
-  - type: "nlp.forte.processors.NLTKSentenceSegmenter"
-  - type: "nlp.forte.processors.NLTKWordTokenizer"
-  - type: "nlp.forte.processors.NLTKPOSTagger"
-  - type: "nlp.forte.processors.CoNLLNERPredictor"
-    hparams:
-      config_path: # Load config from the specified file if any
-      overwrite_configs: # You can still define other configurations here
-        storage_path: "/home/shr/forte/examples/ner/resources.pkl"
-  - type: "nlp.forte.processors.SRLPredictor"
-    hparams:
-      overwrite_configs:
-        storage_path: "/home/hector/models/SRL_model/"
