+++ examples/pipelines/data/sample_text.txt
@@ -1,0 +1,2 @@
+So I was excited to see Journey to the Far Side of the Sun finally get released on an affordable DVD (the previous print had been fetching $100 on eBay - I'm sure those people wish they had their money back - but more about that in a second).
+The plain green Norway spruce is displayed in the gallery's foyer. Wentworth worked as an assistant to sculptor Henry Moore in the late 1960s. His reputation as a sculptor grew in the 1980s.
