+++ forte/data/readers/conll03_reader.py
@@ -99,3 +99,7 @@ class CoNLL03Reader(PackReader):
         doc.close()
 
         yield pack
+
+    @staticmethod
+    def default_hparams():
+        return {}
