+++ forte/multipack_pipeline.py
@@ -1,13 +1,10 @@
 import logging
-from typing import Dict, List, Optional
-
-import yaml
-from texar.torch import HParams
+from typing import Dict, List
 
 from forte.base_pipeline import BasePipeline
 from forte.data.multi_pack import MultiPack
 from forte.data.selector import Selector
-from forte.utils import get_class
+from forte.utils import get_class, create_class_with_kwargs
 
 logger = logging.getLogger(__name__)
 
@@ -66,30 +63,3 @@ class MultiPackPipeline(BasePipeline[MultiPack]):
                 self.add_processor(p, processor_hparams, selector)
 
             self.initialize()
-
-
-def create_class_with_kwargs(
-        class_name: str, class_args: Dict, h_params: Optional[Dict] = None):
-    cls = get_class(class_name)
-    if not class_args:
-        class_args = {}
-    obj = cls(**class_args)
-
-    p_params: Dict = {}
-
-    if h_params is not None and \
-            "config_path" in h_params and \
-            h_params["config_path"] is not None:
-        filebased_hparams = yaml.safe_load(open(h_params["config_path"]))
-    else:
-        filebased_hparams = {}
-    p_params.update(filebased_hparams)
-
-    if h_params is not None:
-        p_params.update(h_params.get("overwrite_configs", {}))
-    default_processor_hparams = cls.default_hparams()
-
-    processor_hparams = HParams(p_params,
-                                default_processor_hparams)
-
-    return obj, processor_hparams
