+++ forte/utils.py
@@ -1,13 +1,18 @@
 """
 Util functions
 """
+from typing import Dict, Optional
 from pydoc import locate
+import yaml
+
+from texar.torch import HParams
 
 __all__ = [
     "get_full_module_name",
     "get_class_name",
     "get_class",
     "get_qual_name",
+    "create_class_with_kwargs"
 ]
 
 
@@ -82,3 +87,30 @@ def get_qual_name(o, lower=False):
         return o.__qualname__.lower()
     else:
         return o.__qualname__
+
+
+def create_class_with_kwargs(
+        class_name: str, class_args: Dict, h_params: Optional[Dict] = None):
+    cls = get_class(class_name)
+    if not class_args:
+        class_args = {}
+    obj = cls(**class_args)
+
+    if h_params is None:
+        h_params = {}
+
+    p_params: Dict = {}
+
+    if "config_path" in h_params and not h_params["config_path"] is None:
+        filebased_hparams = yaml.safe_load(open(h_params["config_path"]))
+    else:
+        filebased_hparams = {}
+    p_params.update(filebased_hparams)
+
+    p_params.update(h_params.get("overwrite_configs", {}))
+    default_processor_hparams = cls.default_hparams()
+
+    processor_hparams = HParams(p_params,
+                                default_processor_hparams)
+
+    return obj, processor_hparams
